requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.12.14
channels==4.2.0
charset-normalizer==3.4.1
crispy-bootstrap4==2024.10
Django==5.1.4
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
networkx==3.4.2
pillow==11.0.0
PyJWT==2.10.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
sqlparse==0.5.2
stripe==11.4.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
